+++ core/src/main/java/com/ensody/reactivestate/WorkQueueCore.kt
@@ -45,13 +45,13 @@ fun <T> CoroutineScope.workQueue() = WorkQueue<T>(this)
 fun <T> State.workQueue() = scope.workQueue<T>()
 
 /** Creates a [WorkQueue] for lambdas taking an argument. You have to manually call `consume()`. */
-fun <T> CoroutineScope.argWorkQueue() = WorkQueue<(T) -> Unit>(this)
+fun <T> CoroutineScope.argWorkQueue() = WorkQueue<suspend (T) -> Unit>(this)
 
 /** Creates a [WorkQueue] for lambdas taking an argument. You have to manually call `consume()`. */
 fun <T> State.argWorkQueue() = scope.argWorkQueue<T>()
 
 /** Creates a [WorkQueue] for lambdas taking a `this` argument. You have to manually call `consume()`. */
-fun <T> CoroutineScope.thisWorkQueue() = WorkQueue<T.() -> Unit>(this)
+fun <T> CoroutineScope.thisWorkQueue() = WorkQueue<suspend T.() -> Unit>(this)
 
 /** Creates a [WorkQueue] for lambdas taking a `this` argument. You have to manually call `consume()`. */
 fun <T> State.thisWorkQueue() = scope.thisWorkQueue<T>()
@@ -159,14 +159,14 @@ class WorkQueue<T>(private val scope: CoroutineScope) : Disposable {
 }
 
 /** Consume work queue, passing [arg] to each lambda. */
-fun <T> WorkQueue<(T) -> Unit>.consume(
+fun <T> WorkQueue<suspend (T) -> Unit>.consume(
     arg: T,
     scope: CoroutineScope,
     workers: Int = 1
 ) = consume(scope, workers = workers) { map { it(arg) } }
 
 /** Consume work queue, conflate lambdas and pass [arg] to each lambda. */
-fun <T> WorkQueue<(T) -> Unit>.conflatedConsume(
+fun <T> WorkQueue<suspend (T) -> Unit>.conflatedConsume(
     arg: T,
     scope: CoroutineScope,
     timeoutMillis: Long = 0,