+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This release introduces no code changes.
 
 ## 0.9
 
-ATTENTION: Due to the core & reactivestate module split you have to also add the core module to your dependencies as described in the [README](README.md).
+ATTENTION: Due to the core & reactivestate module split you have to also add the core module to your dependencies as described in the [installation instructions](https://ensody.github.io/ReactiveState-Kotlin/#installation).
 
 * This release splits the `reactivestate` module into a Kotlin module (`core`) and an Android module (`reactivestate`). The change is backwards-compatible unless you've accessed `BaseAutoRunner` (very unlikely).
 * Added `State` base class for separating business logic from `ViewModel`, making it easier to use in normal unit tests.